+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "masseybot",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "description": "",
   "main": "index.js",
   "scripts": {
@@ -27,10 +27,12 @@
     "express": "^4.17.1",
     "express-handlebars": "^5.3.3",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^6.0.4"
+    "mongoose": "^6.0.4",
+    "piscina": "^3.1.0"
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
-    "@types/jsonwebtoken": "^8.5.5"
+    "@types/jsonwebtoken": "^8.5.5",
+    "@types/workerpool": "^6.1.0"
   }
 }
